+++ crosswrd-app/src/EditLights.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import { Container } from "react-bootstrap";
 
-import { Grid, CellMap } from "./Grid";
+import { Grid, CellMap } from "./Lights";
 import { displayGrid, renderCells } from "./GridDisplay";
 import { WrappedRow, StateSetter, ToggleButton } from "./Helpers";
 import { ExportLights } from "./ExportLights";
